requirements.txt
pyTelegramBotAPI>=4.10.0
pymongo>=4.7
motor>=3.7.1
python-dotenv>=1.1.0
openai>=1.78.1
aiohttp>=3.12.12
cachetools>=5.0
requests>=2.31
tiktoken>=0.3.0